Contoller, service, test fix

+++ service/src/main/java/com/epam/esm/service/impl/TagServiceImpl.java
@@ -9,6 +9,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class TagServiceImpl implements TagService {
@@ -29,8 +30,11 @@ public class TagServiceImpl implements TagService {
     }
 
     @Override
-    public void deleteTag(int id) throws ServiceException {
+    public void deleteTag(Integer id) throws ServiceException {
         try {
+            if (!Optional.ofNullable(id).isPresent() || !isTagExist(id)) {
+                throw new ServiceException("Cant find tag with id:" + id);
+            }
             tagDAO.deleteTag(id);
         } catch (DAOException e) {
             throw new ServiceException("DeleteTag fail.", e);
@@ -47,8 +51,11 @@ public class TagServiceImpl implements TagService {
     }
 
     @Override
-    public Tag getTag(int id) throws ServiceException {
+    public Tag getTag(Integer id) throws ServiceException {
         try {
+            if (!Optional.ofNullable(id).isPresent() || !isTagExist(id)) {
+                throw new ServiceException("Cant find tag with id:" + id);
+            }
             return tagDAO.getTag(id);
         } catch (DAOException e) {
             throw new ServiceException("GetTags fail.", e);
@@ -56,7 +63,7 @@ public class TagServiceImpl implements TagService {
     }
 
     @Override
-    public boolean isTagExist(int id) throws ServiceException {
+    public boolean isTagExist(Integer id) throws ServiceException {
         try {
             return tagDAO.isTagExist(id);
         } catch (DAOException e) {
